vec: use int instead of size_t for type templating (v2)
/home/airlied/devel/compute/conformance-testsuite/util/math_reference.cpp: In function ‘cl::sycl::int2 reference::mad24(cl::sycl::int2, cl::sycl::int2, cl::sycl::int2)’:
/home/airlied/devel/compute/conformance-testsuite/util/math_reference.cpp:871:27: error: no matching function for call to ‘mad24_tmp(cl::sycl::int2&, cl::sycl::int2&, cl::sycl::int2&)’
   return mad24_tmp(x, y, z);
/home/airlied/devel/compute/conformance-testsuite/util/math_reference.cpp:786:26: note: candidate: ‘template<class T, int length> cl::sycl::vec<T, dim> reference::mad24_tmp(cl::sycl::vec<T, dim>, cl::sycl::vec<T, dim>, cl::sycl::vec<T, dim>)’
 cl::sycl::vec<T, length> mad24_tmp(cl::sycl::vec<T, length> x,
                          ^~~~~~~~~
/home/airlied/devel/compute/conformance-testsuite/util/math_reference.cpp:786:26: note:   template argument deduction/substitution failed:
/home/airlied/devel/compute/conformance-testsuite/util/math_reference.cpp:871:27: note:   mismatched types ‘int’ and ‘long unsigned int’
   return mad24_tmp(x, y, z);

Changing this to an int fixes the build for this

v2: fix flatten

+++ include/CL/sycl/vec.hpp
@@ -78,7 +78,7 @@ struct elem {
 
     \todo rename in the specification element_type to value_type
 */
-template <typename DataType, size_t NumElements>
+template <typename DataType, int NumElements>
 class vec : public detail::small_array<DataType,
                                        vec<DataType, NumElements>,
                                        NumElements> {
@@ -114,7 +114,7 @@ private:
       has also a tuple interface :-) (23.3.2.9 Tuple interface to class
       template array [array.tuple])
   */
-  template <typename V, typename Element, size_t s>
+  template <typename V, typename Element, int s>
   static auto flatten(const vec<Element, s> i) {
     static_assert(s <= V::dimension,
                   "The element i will not fit in the vector");
